backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.7.4
pycocotools==2.0.7
opencv-python==4.9.0.80
torch==2.2.2
torchvision==0.17.2
segment-anything-2 @ git+https://github.com/facebookresearch/segment-anything-2.git
